+++ example/src/app.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { McpServerModule } from './mcp-server/mcp.module';
 import { ConfigModule } from '@nestjs/config';
 import { OpenAiModule } from './openai/openai.module';
+import { TelegramModule } from './telegram/telegram.module';
 
 @Module({
   imports: [
@@ -10,6 +11,7 @@ import { OpenAiModule } from './openai/openai.module';
     }),
     OpenAiModule,
     McpServerModule,
+    TelegramModule,
   ],
 })
 export class AppModule {}