+++ example/src/mcp-server/handlers/example.resource.ts
@@ -6,11 +6,11 @@ import {
 import { UseFilters, UseGuards, UseInterceptors } from '@nestjs/common';
 import { TestGuard } from '../lifecicle/test.guard';
 import { TestInterceptor } from '../lifecicle/test.interceptor';
-import { TestFilter } from '../lifecicle/test.filter';
+import { AuthFilter } from '../lifecicle/test.filter';
 
 @UseGuards(TestGuard)
 @UseInterceptors(TestInterceptor)
-@UseFilters(TestFilter)
+@UseFilters(AuthFilter)
 @McpResource()
 export class ExampleResource implements IMcpResource<{ userId: string }> {
   name = 'users.get';