+++ example/src/mcp-server/lifecicle/test.filter.ts
@@ -1,13 +1,31 @@
-import { McpUnauthorizedException } from '@muzikanto/nestjs-mcp';
+import {
+  McpBadRequestException,
+  McpUnauthorizedException,
+} from '@muzikanto/nestjs-mcp';
 import { ExceptionFilter, Catch, ArgumentsHost } from '@nestjs/common';
 
 @Catch(McpUnauthorizedException)
-export class TestFilter implements ExceptionFilter {
+export class AuthFilter implements ExceptionFilter {
   catch(exception: unknown, host: ArgumentsHost) {
     return {
       isError: true,
       messages: [
-        { type: 'text', text: (exception as Error).message || 'No access' },
+        { role: 'user', content: (exception as Error).message || 'No access' },
+      ],
+    };
+  }
+}
+
+@Catch(McpBadRequestException)
+export class BadRequestFilter implements ExceptionFilter {
+  catch(exception: unknown, host: ArgumentsHost) {
+    return {
+      isError: true,
+      messages: [
+        {
+          role: 'user',
+          content: (exception as Error).message || 'invalid arguments',
+        },
       ],
     };
   }